Remove userHasVideo model.
And replace it with a user_id field in the video model.
If we *really* want to have sets of videos that we provide to users in
the future, we can assign them to users via copies in the videos table.
I'm guessing that videos will be pretty low volume as well as low overlap.

Closes #53